created package deck, deck/jobs, which contains logic for job and config agents and pod logs, moved associated logic out of the main package and refactored deck accordingly.
fully tested and refactored JobAgent to be outside main package
