+++ prow/BUILD.bazel
@@ -81,6 +81,7 @@ filegroup(
         "//prow/commentpruner:all-srcs",
         "//prow/config:all-srcs",
         "//prow/cron:all-srcs",
+        "//prow/deck/jobs:all-srcs",
         "//prow/entrypoint:all-srcs",
         "//prow/errorutil:all-srcs",
         "//prow/external-plugins/cherrypicker:all-srcs",
